データ削除時の戻り先を修正

+++ app/Http/Controllers/Asset/AssetsController.php
@@ -106,11 +106,9 @@ class AssetsController extends Controller
             DB::commit();
 
             return redirect()->route('assets.dashboard')->with('success-message', __('create_success_message'));
-
         } catch (\Exception $e) {
             DB::rollBack();
             Log::error(__('create_error_log'), [
-                'stack_trace' => $e->getTraceAsString(),
                 'error_message' => $e->getMessage(),
                 'method' => request()->method(),
                 'parameters' => request()->all(),
@@ -169,7 +167,6 @@ class AssetsController extends Controller
             } catch (\Exception $e) {
                 DB::rollBack();
                 Log::error(__('update_error_log'), [
-                    'stack_trace' => $e->getTraceAsString(),
                     'error_message' => $e->getMessage(),
                     'method' => request()->method(),
                     'parameters' => request()->all(),
@@ -190,7 +187,6 @@ class AssetsController extends Controller
             } catch (\Exception $e) {
                 DB::rollBack();
                 Log::error(__('new_error_log'), [
-                    'stack_trace' => $e->getTraceAsString(),
                     'error_message' => $e->getMessage(),
                     'method' => request()->method(),
                     'parameters' => request()->all(),
@@ -206,21 +202,26 @@ class AssetsController extends Controller
      * @return \Illuminate\Http\RedirectResponse リダイレクトレスポンスを返す
      * @throws \Exception 資産削除時に例外が発生した場合
      */
-    public function destroy(string $id)
+    public function destroy(string $id, Request $request)
     {
 
+        $parsedUrl = $this->assetService->generateRedirectUrl($request);
+
         $asset = Asset::find($id);
+        if (empty($asset)) {
+            Log::error(__('destroy_error_message'));
+            return back()->withInput()->with('error-message', __('delete_error_message'));
+        }
 
         try {
             DB::beginTransaction();
             $asset->delete();
             DB::commit();
 
-            return redirect()->route('assets.index')->with('success-message', __('delete_success_message'));
+            return redirect($parsedUrl)->with('success-message', __('delete_success_message'));
         } catch (\Exception $e) {
             DB::rollBack();
             Log::error(__('delete_error_log'), [
-                'stack_trace' => $e->getTraceAsString(),
                 'error_message' => $e->getMessage(),
                 'method' => request()->method(),
                 'parameters' => request()->all(),
